Encapsulate Late Pledge Apple Pay logic into it's own ApplePayCheckoutService
